requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dataclasses-json==0.5.7
filelock==3.11.0
Flask==2.0.2
frozenlist==1.3.3
gpt-index==0.4.0
greenlet==2.0.2
huggingface-hub==0.13.4
idna==3.4
importlib-metadata==6.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.79
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.24.2
openai==0.27.4
packaging==23.0
pandas==2.0.0
pathspec==0.11.1
platformdirs==3.2.0
pydantic==1.10.7
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.47
tenacity==8.1.0
tiktoken==0.3.3
tokenizers==0.13.3
tomli==2.0.1
tqdm==4.65.0
transformers==4.27.4
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
